refactor: Rename getReadByText to listUsers
